Rename Headding component directory to Heading
